Finally got the number without a return
i need to understand return better. But might have to play with main to get some understanding of the list. Think i will have to dataframe it. But not sure. All this to find an original list for the end when I place amps in correct spots along with 0 for spares. and i have not completed figuring out the lineWatts of a circuit